mock_interview_platform/requirements.txt
agent-framework-azure-ai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
google-generativeai>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0